Method for read and cw

+++ IZT6ZK/Commands/DeleteQuestionCommand.cs
@@ -12,7 +12,7 @@ internal class DeleteQuestionCommand : ICommands
     {
         var dbManager = new DbManager();
         string? inputQuestionId;
-        Console.WriteLine("\nWrite 'quit' if you want to quit\n");
+        Console.WriteLine("\nWrite 'quit' if you want to quit.");
 
         while (true)
         {
@@ -23,7 +23,7 @@ internal class DeleteQuestionCommand : ICommands
                 Console.WriteLine($"{allQuestion.QuestionId}: {allQuestion.Question}");
             }
 
-            Console.WriteLine("Write the question's id, you want to delete: ");
+            Console.WriteLine("\nWrite the question's id, you want to delete: ");
             inputQuestionId = Console.ReadLine();
 
             if (string.IsNullOrEmpty(inputQuestionId) || string.IsNullOrWhiteSpace(inputQuestionId))
@@ -35,20 +35,35 @@ internal class DeleteQuestionCommand : ICommands
 
             if (inputQuestionId == "quit")
             {
-                Console.WriteLine("You quitted \n");
+                Console.WriteLine("You quitted! \n");
                 break;
             }
             int.TryParse(inputQuestionId, out var questionId);
             var questionEntity = dbManager.SelectQuestion(questionId);
             if (questionEntity != null)
             {
-                dbManager.DeleteQuestion(questionEntity);
-                Console.WriteLine("Congratulations, you deleted the question!\n");
-                break;
+                Console.WriteLine("Are you sure? Yes or No");
+                var yesOrNo = Console.ReadLine();
+                yesOrNo = yesOrNo.Trim().ToLower();
+                if (yesOrNo == "no")
+                {
+                    Console.WriteLine("You didn't delete the question!");
+                    continue;
+                }
+                else if (yesOrNo == "yes")
+                {
+                    dbManager.DeleteQuestion(questionEntity);
+                    Console.WriteLine("\nCongratulations, you deleted the question!\n");
+                    break;
+                }
+                else 
+                {
+                    Console.WriteLine("Please write 'yes' or 'no'!");
+                    continue;
+                }
             }
             Console.WriteLine("Please write an existing question id!");
             continue;
-
         }
     }
 }
